fpga-cfg: update manager's usb id parsing
Since commit ID 03a0ee5081 the FPP manager name contains
board address specifier: "right" or "left" for board revB
or "single" for board revA. Update the driver to parse the
address specifier and usb id properly.

We have to create a unique configuration interface path
depending on the detected address specifier to differentiate
between the left and right PRAX boards. This needs to be
addressed in a separate patch.

Signed-off-by: Anatolij Gustschin <agust@denx.de>