fpga-cfg: rework for unique interface path for FPP/SPI
With RevB PRAX board the board address string is encoded in the fpga
manager name. Use this string to create unique path for configuration
interface. For RevA board without address_sel logic the value
"fpp_single.N" is used as the directory name for configuration
interface, for RevB board the value "fpp_right.N" or "fpp_left.N"
is used, e.g.:

  PRAX RevA:
    /sys/kernel/debug/fpga_cfg/fpp_single.1

  PRAX RevB:
    /sys/kernel/debug/fpga_cfg/fpp_right.0
    /sys/kernel/debug/fpga_cfg/fpp_left.1

For SPI configuration interfaces the SPI slave name is encoded
in the fpga manager name. The created directory to the configuration
interface looks like "spi_spiN.M", where N is the SPI bus number and
M is the chip select number of assigned to the SPI slave. E.g.:

  /sys/kernel/debug/fpga_cfg/spi_spi0.1/

Signed-off-by: Anatolij Gustschin <agust@denx.de>