fpga-cfg: request xilinx-spi module when loading
On ARM Amira camera we use xilinx-spi driver as FPGA manager for loading
Xilinx Spartan-6 FPGA. The fpga-cfg driver also creates the configuration
interface for it and needs to find this FPGA manager.

Request appropriate driver module before searching for existing FPGA
managers. This will ensure that the config interface is created (provided
that the xilinx-spi driver module is enabled in the kernel config and is
installed in the rootfs).

Signed-off-by: Anatolij Gustschin <agust@denx.de>