fpga-cfg: Pass the module parameters per device and not per driver
This patch changes the passing of the module parameters. Until now these
parameters were passed to the specified driver as module parameters
while driver loading. This has the disadvantage that different parameters
can't be used with identical drivers. This patch now changes this by
passing the module parameters via a newly created platform_data buffer,
which is then passed to the device in pci_device_driver_bind(). This
string contains a plain copy of the parameters defined in the config
file "mfd-driver-param" key. The driver then can parse this string
with these parameters on a per-device basis. With this change these
parameters are device specific and we can support different parameters
for different FPGA's, even then the same device driver is used.

Please note that the module parameters are now not separated by ',' any
more. Please separate them simply via a space instead, like this:

  mfd-driver-param = "mfd_bar_nr=1 mfd_bar_offs=0";

Signed-off-by: Stefan Roese <sr@denx.de>